CORE-7627 Refactor *FileUploadDialog
SimpleFileUploadDialog and EdamFileUploadDialog now extend the abstract class AbstractFileUploadDialog.  All 3 are now located in the commons area.
The abstract class manages the UI elements.  The subclasses only have to manage handling the upload and handling the response afterwards.
