--- ui/de-lib/src/main/java/org/iplantc/de/admin/desktop/client/ontologies/views/dialogs/EdamUploadPanel.ui.xml
@@ -1,69 +1,0 @@
-<!DOCTYPE ui:UiBinder SYSTEM "http://dl.google.com/gwt/DTD/xhtml.ent">
-<ui:UiBinder xmlns:ui="urn:ui:com.google.gwt.uibinder"
-             xmlns:g="urn:import:com.google.gwt.user.client.ui"
-             xmlns:con="urn:import:com.sencha.gxt.widget.core.client.container"
-             xmlns:form="urn:import:org.iplantc.de.commons.client.widgets"
-             xmlns:gxtForm="urn:import:com.sencha.gxt.widget.core.client.form"
-             xmlns:button="urn:import:com.sencha.gxt.widget.core.client.button"
-             xmlns:gxt="urn:import:com.sencha.gxt.widget.core.client">
-
-    <ui:with field="appearance"
-             type="org.iplantc.de.admin.desktop.client.ontologies.OntologiesView.OntologiesViewAppearance"/>
-
-    <ui:with field="sideMargins"
-             type="com.sencha.gxt.core.client.util.Margins">
-        <ui:attributes top="4"
-                       right="4"
-                       bottom="4"
-                       left="4"/>
-    </ui:with>
-
-    <ui:with field="centerMargins"
-             type="com.sencha.gxt.core.client.util.Margins">
-        <ui:attributes top="4"
-                       right="0"
-                       bottom="4"
-                       left="0"/>
-    </ui:with>
-
-    <ui:with field="centerLayoutData"
-             type="com.sencha.gxt.widget.core.client.container.HorizontalLayoutContainer.HorizontalLayoutData">
-        <ui:attributes width="1"
-                       height="1"
-                       margins="{centerMargins}"/>
-    </ui:with>
-
-    <ui:with field="sideLayoutData"
-             type="com.sencha.gxt.widget.core.client.container.HorizontalLayoutContainer.HorizontalLayoutData">
-        <ui:attributes width="-1"
-                       height="1"
-                       margins="{sideMargins}"/>
-    </ui:with>
-
-    <con:VerticalLayoutContainer>
-        <g:HTML text="{appearance.fileUploadMaxSizeWarning}"/>
-        <g:HTML ui:field="htmlDestText"/>
-
-        <!-- One -->
-        <gxtForm:FormPanel ui:field="form"
-        borders="false">
-            <con:HorizontalLayoutContainer>
-                <con:child layoutData="{sideLayoutData}">
-                    <button:TextButton ui:field="btn"
-                                       toolTip="{appearance.reset}"
-                                       icon="{appearance.arrowUndoIcon}"
-                                       debugId="idBtnReset0"/>
-                </con:child>
-                <con:child layoutData="{centerLayoutData}">
-                    <form:IPCFileUploadField ui:field="fuf"
-                                             debugId="idFileUpld0"
-                                             name="file"/>
-                </con:child>
-                <con:child layoutData="{sideLayoutData}">
-                    <gxt:Status ui:field="status"/>
-                </con:child>
-            </con:HorizontalLayoutContainer>
-        </gxtForm:FormPanel>
-    </con:VerticalLayoutContainer>
-</ui:UiBinder>
-
