+++ ui/de-lib/src/main/java/org/iplantc/de/diskResource/client/views/dialogs/BulkMetadataPanel.ui.xml
@@ -8,8 +8,8 @@
 	xmlns:gxt="urn:import:com.sencha.gxt.widget.core.client">
 
 	<ui:with field="appearance"
-		type="org.iplantc.de.diskResource.client.views.dialogs.SimpleFileUploadDialog.SimpleFileUploadDialogAppearance" />
-		
+			 type="org.iplantc.de.commons.client.views.dialogs.AbstractFileUploadDialog.AbstractFileUploadDialogAppearance"/>
+
 		<ui:with field="bulkapp"
         type="org.iplantc.de.diskResource.client.views.dialogs.BulkMetadataDialog.Appearance" />
