+++ ui/de-lib/src/main/java/org/iplantc/de/theme/base/client/diskResource/DiskResourceMessages.properties
@@ -86,4 +86,4 @@ bulkMetadataError = Unable to apply metadata!
 overWiteMetadata = Some metadata already exists. Do you want to overwrite them ?
 doiRequestFail =  DOI Request failed!
 doiRequestSuccess = Your request for DOI has been submitted! You will receive notification when there are updates to this request.
-
+fileUploadsSuccess = The file(s) {0,list} were successfully uploaded.
